move code and docs
